LayoutTests/media/modern-media-controls/tracks-panel/tracks-panel-controls-bar-remains-visible-after-clicking-over-it-expected.txt

Clicking on the controls bar to dismiss the tracks panel should not hide the controls bar, only the track panel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


mediaControls.showTracksPanel()
PASS mediaControls.tracksPanel.element.parentNode became mediaControls.element

We've waited long enough so that the controls bar should be faded should the tracks panel not be visible
PASS mediaControls.controlsBar.faded is false

Click on the controls bar to dismiss the panel, the controls bar should remain visible
PASS mediaControls.controlsBar.faded is false
PASS mediaControls.tracksPanel.presented is false

PASS successfullyParsed is true

TEST COMPLETE
